feat: refactored basic services
