+++ greyhoundDashboard/services/basic_services/__init__.py
@@ -1,97 +1,23 @@
-from . import *
-from services.service_base import ServiceBase, EnhancedServiceBase
+# module/__init__.py
+"""
+Expose every class defined in any sub-module whose name ends with *Service*
+(or pick whatever rule you need) so callers can simply::
 
-class ServiceProxmox(ServiceBase):
+    from module import MyService
+"""
+import pkgutil
+import importlib
+import inspect
 
-    """
-    🖥️ < Virtualization!
-    """
-    id:str = 'proxmox'
-    name:str = 'Proxmox'
-    description:str = 'An open-source virtualization management platform for running and managing virtual machines and containers using KVM and LXC technologies.'
-    tags: list[str] = ['virtualization', 'infrastructure', 'hypervisor', 'containers', 'kvm', 'self-hosted']
+__all__: list[str] = []          # what `from module import *` will export
+_pkg_name = __name__             # e.g. "module"
 
-    def get(self) -> ServiceBase:
+for modinfo in pkgutil.walk_packages(__path__, prefix=f"{_pkg_name}."):
+    # 1. import the sub-module (e.g. module.service_test)
+    mod = importlib.import_module(modinfo.name)
 
-        super().get()
-
-        return self
-
-
-class ServiceOpenWebUI(ServiceBase):
-
-    """
-    🤖 < AI!
-    """
-    id:str = 'openwebui'
-    name:str = 'Open WebUI'
-    description:str = 'A sleek, self-hosted interface for interacting with large language models like ChatGPT, supporting local backends such as Ollama and LM Studio.'
-    tags: list[str] = ['AI', 'LLM', 'chatbot', 'self-hosted', 'open-source']
-
-    def get(self) -> ServiceBase:
-
-        super().get()
-
-        return self
-    
-
-class ServiceJenkins(ServiceBase):
-    """
-    🔧 Jenkins CI/CD
-    """
-    id: str = 'jenkins'
-    name: str = 'Jenkins'
-    description: str = 'An open-source automation server that enables developers to build, test, and deploy their software with ease.'
-    tags: list[str] = ['CI/CD', 'DevOps', 'automation', 'self-hosted', 'open-source']
-
-    def get(self) -> ServiceBase:
-        super().get()
-        return self
-    
-
-class ServiceGrafana(ServiceBase):
-    """
-    📊 Grafana
-    """
-    id: str = 'grafana'
-    name: str = 'Grafana'
-    description: str = 'An open-source platform for monitoring and observability, offering beautiful dashboards and integrations with Prometheus, Loki, and other data sources.'
-    tags: list[str] = ['monitoring', 'dashboards', 'observability', 'self-hosted', 'open-source']
-
-    def get(self) -> ServiceBase:
-        super().get()
-        return self
-
-
-class ServicePihole(ServiceBase):
-    """
-    🛡️ Pi-hole
-    """
-    id: str = 'pihole'
-    name: str = 'Pi-hole'
-    description: str = 'A network-wide ad blocker that acts as a DNS sinkhole to block ads, trackers, and malicious domains for all devices on your network.'
-    tags: list[str] = ['network', 'ad-blocking', 'dns', 'privacy', 'self-hosted', 'open-source']
-
-    def get(self) -> ServiceBase:
-        super().get()
-        return self
-
-
-class ServiceOllama(ServiceBase):
-    """
-    🦙 Ollama
-    """
-    id: str = 'ollama'
-    name: str = 'Ollama'
-    description: str = 'Run large language models locally with ease. Ollama provides a simple interface and support for models like LLaMA, Mistral, and more.'
-    tags: list[str] = ['LLM', 'AI', 'local', 'self-hosted', 'open-source']
-
-    def get(self) -> ServiceBase:
-        super().get()
-        return self
-
-
-__all__ = [
-    name for name, obj in globals().items()
-    if name.startswith('Service') and isinstance(obj, type)
-]
+    # 2. promote matching classes to the package namespace
+    for name, obj in vars(mod).items():
+        if inspect.isclass(obj) and name.startswith("Service"):
+            globals()[name] = obj   # expose at package level
+            __all__.append(name)    # re-export
